+++ Tests/test_content.py
@@ -1,21 +1,23 @@
 import re
 import sys
 import json
-import string
-import random
 import argparse
 import requests
 import subprocess
+import urllib3
 from time import sleep
 from datetime import datetime
 
-import demisto
+import demisto_client.demisto_api
 from slackclient import SlackClient
 
 from Tests.test_integration import test_integration, disable_all_integrations
 from Tests.mock_server import MITMProxy, AMIConnection
 from Tests.test_utils import print_color, print_error, print_warning, LOG_COLORS, str2bool, server_version_compare
 from Tests.scripts.constants import RUN_ALL_TESTS_FORMAT, FILTER_CONF, PB_Status
+
+# Disable insecure warnings
+urllib3.disable_warnings()
 
 SERVER_URL = "https://{}"
 INTEGRATIONS_CONF = "./Tests/integrations_file.txt"
@@ -151,7 +153,7 @@ def send_slack_message(slack, chanel, text, user_name, as_user):
 def run_test_logic(c, failed_playbooks, integrations, playbook_id, succeed_playbooks, test_message, test_options, slack,
                    circle_ci, build_number, server_url, build_name, is_mock_run=False):
     status, inc_id = test_integration(c, integrations, playbook_id, test_options, is_mock_run)
-
+    # c.api_client.pool.close()
     if status == PB_Status.COMPLETED:
         print_color('PASS: {} succeed'.format(test_message), LOG_COLORS.GREEN)
         succeed_playbooks.append(playbook_id)
@@ -227,19 +229,20 @@ def mock_run(c, proxy, failed_playbooks, integrations, playbook_id, succeed_play
     print('------ Test {} end ------\n'.format(test_message))
 
 
-def run_test(c, proxy, failed_playbooks, integrations, unmockable_integrations, playbook_id, succeed_playbooks,
+def run_test(demisto_api_key, proxy, failed_playbooks, integrations, unmockable_integrations, playbook_id, succeed_playbooks,
              test_message, test_options, slack, circle_ci, build_number, server_url, build_name, is_ami=True):
     start_message = '------ Test %s start ------' % (test_message,)
+    client = demisto_client.configure(base_url=server_url, api_key=demisto_api_key, verify_ssl=False)
 
     if not is_ami or (not integrations or has_unmockable_integration(integrations, unmockable_integrations)):
         print(start_message + ' (Mock: Disabled)')
-        run_test_logic(c, failed_playbooks, integrations, playbook_id, succeed_playbooks, test_message, test_options,
+        run_test_logic(client, failed_playbooks, integrations, playbook_id, succeed_playbooks, test_message, test_options,
                        slack, circle_ci, build_number, server_url, build_name)
         print('------ Test %s end ------\n' % (test_message,))
 
         return
 
-    mock_run(c, proxy, failed_playbooks, integrations, playbook_id, succeed_playbooks,
+    mock_run(client, proxy, failed_playbooks, integrations, playbook_id, succeed_playbooks,
              test_message, test_options, slack, circle_ci, build_number, server_url, build_name, start_message)
 
 
@@ -373,16 +376,6 @@ def extract_filtered_tests():
     return filtered_tests, is_filter_configured, run_all
 
 
-def generate_demisto_api_key(c):
-    demisto_api_key = ''.join(random.choice(string.ascii_letters + string.digits) for _ in range(32))
-    apikey_json = {
-        'name': 'test_apikey',
-        'apikey': demisto_api_key
-    }
-    c.req('POST', '/apikeys', apikey_json)
-    return demisto_api_key
-
-
 def load_conf_files(conf_path, secret_conf_path):
     with open(conf_path) as data_file:
         conf = json.load(data_file)
@@ -414,11 +407,11 @@ def organize_tests(tests, unmockable_integrations, skipped_integrations_conf, ni
     return mock_tests, mockless_tests
 
 
-def run_test_scenario(t, c, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
+def run_test_scenario(t, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
                       skipped_integrations_conf, skipped_integration, is_nightly, run_all_tests, is_filter_configured,
-                      filtered_tests, skipped_tests, demisto_api_key, secret_params, failed_playbooks,
+                      filtered_tests, skipped_tests, secret_params, failed_playbooks,
                       unmockable_integrations, succeed_playbooks, slack, circle_ci, build_number, server, build_name,
-                      server_numeric_version, is_ami=True):
+                      server_numeric_version, demisto_api_key, is_ami=True):
     playbook_id = t['playbookID']
     nightly_test = t.get('nightly', False)
     integrations_conf = t.get('integrations', [])
@@ -491,12 +484,13 @@ def run_test_scenario(t, c, proxy, default_test_timeout, skipped_tests_conf, nig
         text = stdout if not stderr else stderr
         send_slack_message(slack, SLACK_MEM_CHANNEL_ID, text, 'Content CircleCI', 'False')
 
-    run_test(c, proxy, failed_playbooks, integrations, unmockable_integrations, playbook_id,
+    run_test(demisto_api_key, proxy, failed_playbooks, integrations, unmockable_integrations, playbook_id,
              succeed_playbooks, test_message, test_options, slack, circle_ci,
              build_number, server, build_name, is_ami)
 
 
-def restart_demisto_service(ami, c):
+def restart_demisto_service(ami, demisto_api_key, server):
+    client = demisto_client.configure(base_url=server, api_key=demisto_api_key, verify_ssl=False)
     ami.check_call(['sudo', 'service', 'demisto', 'restart'])
     exit_code = 1
     for _ in range(0, SERVICE_RESTART_TIMEOUT, SERVICE_RESTART_POLLING_INTERVAL):
@@ -504,10 +498,10 @@ def restart_demisto_service(ami, c):
         if exit_code != 0:
             exit_code = ami.call(['/usr/sbin/service', 'demisto', 'status', '--lines', '0'])
         if exit_code == 0:
-            print("{}: Checking login to the server...".format(datetime.now()))
+            print("{}: Checking login to the server... ".format(datetime.now()))
             try:
-                res = c.Login()
-                if res.status_code == 200:
+                res = demisto_client.generic_request_func(self=client, path='/health', method='GET')
+                if int(res[1]) == 200:
                     return
                 else:
                     print("Failed verifying login (will retry). status: {}. text: {}".format(res.status_code, res.text))
@@ -521,8 +515,6 @@ def execute_testing(server, server_ip, server_version, server_numeric_version, i
     print("Executing tests with the server {} - and the server ip {}".format(server, server_ip))
 
     options = options_handler()
-    username = options.user
-    password = options.password
     conf_path = options.conf
     secret_conf_path = options.secret
     is_nightly = options.nightly
@@ -532,25 +524,14 @@ def execute_testing(server, server_ip, server_version, server_numeric_version, i
     build_number = options.buildNumber
     build_name = options.buildName
 
-    if not (username and password and server):
-        print_error('You must provide server user & password arguments')
-        sys.exit(1)
-
-    c = demisto.DemistoClient(None, server, username, password)
-    res = c.Login()
-    if res.status_code != 200:
-        print_error("Login has failed with status code " + str(res.status_code))
-        sys.exit(1)
-
-    demisto_api_key = generate_demisto_api_key(c)
-
     conf, secret_conf = load_conf_files(conf_path, secret_conf_path)
+    demisto_api_key = secret_conf.get('temp_apikey')
 
     default_test_timeout = conf.get('testTimeout', 30)
 
     tests = conf['tests']
     skipped_tests_conf = conf['skipped_tests']
-    nightly_integrations = conf['nigthly_integrations']
+    nightly_integrations = conf['nightly_integrations']
     skipped_integrations_conf = conf['skipped_integrations']
     unmockable_integrations = conf['unmockable_integrations']
 
@@ -568,14 +549,14 @@ def execute_testing(server, server_ip, server_version, server_numeric_version, i
     if is_ami:
         ami = AMIConnection(server_ip)
         ami.clone_mock_data()
-        proxy = MITMProxy(c, server_ip)
+        proxy = MITMProxy(server_ip)
 
     failed_playbooks = []
     succeed_playbooks = []
     skipped_tests = set([])
     skipped_integration = set([])
 
-    disable_all_integrations(c)
+    disable_all_integrations(demisto_api_key, server)
 
     if is_ami:
         # move all mock tests to the top of the list
@@ -591,28 +572,27 @@ def execute_testing(server, server_ip, server_version, server_numeric_version, i
                            'Content CircleCI', 'False')
     # first run the mock tests to avoid mockless side effects in container
     if is_ami and mock_tests:
-        proxy.configure_proxy_in_demisto(proxy.ami.docker_ip + ':' + proxy.PROXY_PORT)
+        proxy.configure_proxy_in_demisto(demisto_api_key, server, proxy.ami.docker_ip + ':' + proxy.PROXY_PORT)
         for t in mock_tests:
-            run_test_scenario(t, c, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
+            run_test_scenario(t, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
                               skipped_integrations_conf, skipped_integration, is_nightly, run_all_tests,
                               is_filter_configured,
-                              filtered_tests, skipped_tests, demisto_api_key, secret_params, failed_playbooks,
+                              filtered_tests, skipped_tests, secret_params, failed_playbooks,
                               unmockable_integrations, succeed_playbooks, slack, circle_ci, build_number, server,
-                              build_name, server_numeric_version)
+                              build_name, server_numeric_version, demisto_api_key)
 
         print("\nRunning mock-disabled tests")
-        proxy.configure_proxy_in_demisto('')
+        proxy.configure_proxy_in_demisto(demisto_api_key, server, '')
         print("Restarting demisto service")
-        restart_demisto_service(ami, c)
+        restart_demisto_service(ami, demisto_api_key, server)
         print("Demisto service restarted\n")
-
     for t in mockless_tests:
-        run_test_scenario(t, c, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
+        run_test_scenario(t, proxy, default_test_timeout, skipped_tests_conf, nightly_integrations,
                           skipped_integrations_conf, skipped_integration, is_nightly, run_all_tests,
                           is_filter_configured,
-                          filtered_tests, skipped_tests, demisto_api_key, secret_params, failed_playbooks,
+                          filtered_tests, skipped_tests, secret_params, failed_playbooks,
                           unmockable_integrations, succeed_playbooks, slack, circle_ci, build_number, server,
-                          build_name, server_numeric_version, is_ami)
+                          build_name, server_numeric_version, demisto_api_key, is_ami)
 
     print_test_summary(succeed_playbooks, failed_playbooks, skipped_tests, skipped_integration, unmockable_integrations,
                        proxy, is_ami)
@@ -624,11 +604,12 @@ def execute_testing(server, server_ip, server_version, server_numeric_version, i
         ami.upload_mock_files(build_name, build_number)
 
     if len(failed_playbooks):
-        file_path = "./Tests/is_build_failed_{}.txt".format(server_version.replace(' ', ''))
-        with open(file_path, "w") as is_build_failed_file:
-            is_build_failed_file.write('Build failed')
-
+        print("Some tests have failed. Not destroying instances.")
         sys.exit(1)
+    else:
+        file_path = "./Tests/is_build_passed_{}.txt".format(server_version.replace(' ', ''))
+        with open(file_path, "w") as is_build_passed_file:
+            is_build_passed_file.write('Build passed')
 
 
 def main():