+++ Integrations/Palo_Alto_Networks_WildFire_v2/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+Added a validation to the **server** parameter.
 
 ## [19.10.2] - 2019-10-29
 Fixed an issue in which testing the integration instance failed.